+++ ios/MullvadVPN/SelectLocationViewController.swift
@@ -65,6 +65,20 @@ class SelectLocationViewController: UIViewController, UITableViewDelegate {
             comment: ""
         )
 
+        let searchBar = UISearchBar()
+        searchBar.translatesAutoresizingMaskIntoConstraints = false
+        searchBar.searchBarStyle = .minimal
+        searchBar.barStyle = .black
+        searchBar.tintColor = .lightGray
+        searchBar.delegate = self
+        searchBar.placeholder = NSLocalizedString(
+            "SEARCH_PLACEHOLDER",
+            tableName: "SelectLocation",
+            value: "Search Location",
+            comment: ""
+        )
+        view.addSubview(searchBar)
+
         let tableView = UITableView(frame: view.bounds, style: .plain)
         tableView.translatesAutoresizingMaskIntoConstraints = false
         tableView.backgroundColor = .clear
@@ -72,6 +86,10 @@ class SelectLocationViewController: UIViewController, UITableViewDelegate {
         tableView.separatorInset = .zero
         tableView.estimatedRowHeight = 53
         tableView.indicatorStyle = .white
+        tableView.keyboardDismissMode = .onDrag
+        if #available(iOS 13.0, *) {
+            tableView.automaticallyAdjustsScrollIndicatorInsets = false
+        }
 
         tableView.register(
             SelectLocationCell.self,
@@ -80,7 +98,7 @@ class SelectLocationViewController: UIViewController, UITableViewDelegate {
 
         self.tableView = tableView
 
-        view.accessibilityElements = [tableHeaderFooterView, tableView]
+        view.accessibilityElements = [tableHeaderFooterView, searchBar, tableView]
         view.backgroundColor = .secondaryColor
         view.addSubview(tableView)
 
@@ -90,7 +108,9 @@ class SelectLocationViewController: UIViewController, UITableViewDelegate {
         dataSource = LocationDataSource(
             tableView: tableView,
             cellProvider: { [weak self] tableView, indexPath, item -> UITableViewCell? in
-                guard let self = self else { return nil }
+                guard let self = self else {
+                    return nil
+                }
 
                 let cell = tableView.dequeueReusableCell(
                     withIdentifier: Self.cellReuseIdentifier,
@@ -103,8 +123,13 @@ class SelectLocationViewController: UIViewController, UITableViewDelegate {
                 cell.locationLabel.text = item.displayName
                 cell.showsCollapseControl = item.isCollapsible
                 cell.isExpanded = item.showsChildren
+                cell.isPinned = item.isPinned
+                cell.pinButton.isHidden = !item.isPinnable
                 cell.didCollapseHandler = { [weak self] cell in
                     self?.collapseCell(cell)
+                }
+                cell.didTapPinHandler = { [weak self] cell in
+                    self?.dataSource?.togglePin(item.location)
                 }
 
                 return cell
@@ -116,20 +141,23 @@ class SelectLocationViewController: UIViewController, UITableViewDelegate {
 
         tableHeaderFooterViewTopConstraints = [
             tableHeaderFooterView.topAnchor.constraint(equalTo: view.topAnchor),
-            tableView.topAnchor.constraint(equalTo: tableHeaderFooterView.bottomAnchor),
-            tableView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
+            searchBar.topAnchor.constraint(equalTo: tableHeaderFooterView.bottomAnchor),
+            tableView.bottomAnchor.constraint(equalTo: view.bottomAnchor)
         ]
         tableHeaderFooterViewBottomConstraints = [
             tableHeaderFooterView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
-            tableView.topAnchor.constraint(equalTo: view.topAnchor),
-            tableView.bottomAnchor.constraint(equalTo: tableHeaderFooterView.topAnchor),
+            searchBar.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
+            tableView.bottomAnchor.constraint(equalTo: tableHeaderFooterView.topAnchor)
         ]
 
         NSLayoutConstraint.activate([
             tableHeaderFooterView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
             tableHeaderFooterView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
+            searchBar.leadingAnchor.constraint(equalTo: view.leadingAnchor),
+            searchBar.trailingAnchor.constraint(equalTo: view.trailingAnchor),
+            tableView.topAnchor.constraint(equalTo: searchBar.bottomAnchor),
             tableView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            tableView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
+            tableView.trailingAnchor.constraint(equalTo: view.trailingAnchor)
         ])
         setTableHeaderFooterConstraints()
 
@@ -155,8 +183,7 @@ class SelectLocationViewController: UIViewController, UITableViewDelegate {
         showHeaderViewAtTheBottom = presentingViewController == nil
 
         if let indexPath = dataSource?.indexPathForSelectedRelay(),
-           scrollToSelectedRelayOnViewWillAppear, !isViewAppeared
-        {
+           scrollToSelectedRelayOnViewWillAppear, !isViewAppeared {
             tableView?.scrollToRow(at: indexPath, at: .middle, animated: false)
         }
     }
@@ -192,6 +219,7 @@ class SelectLocationViewController: UIViewController, UITableViewDelegate {
         super.viewDidLayoutSubviews()
 
         updateTableHeaderTopLayoutMargin()
+        adjustTableViewScrollIndicatorInsets()
     }
 
     // MARK: - UITableViewDelegate
@@ -210,14 +238,15 @@ class SelectLocationViewController: UIViewController, UITableViewDelegate {
         forRowAt indexPath: IndexPath
     ) {
         if let item = dataSource?.item(for: indexPath),
-           item.location == dataSource?.selectedRelayLocation
-        {
+           item.location == dataSource?.selectedRelayLocation {
             cell.setSelected(true, animated: false)
         }
     }
 
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        guard let item = dataSource?.item(for: indexPath) else { return }
+        guard let item = dataSource?.item(for: indexPath) else {
+            return
+        }
 
         dataSource?.setSelectedRelayLocation(
             item.location,
@@ -277,8 +306,7 @@ class SelectLocationViewController: UIViewController, UITableViewDelegate {
         // When contained within the navigation controller, we want the distance between the navigation title
         // and the table header label to be exactly 24pt.
         if let navigationBar = navigationController?.navigationBar as? CustomNavigationBar,
-           !showHeaderViewAtTheBottom
-        {
+           !showHeaderViewAtTheBottom {
             tableHeaderFooterView.topLayoutMarginAdjustmentForNavigationBarTitle = navigationBar
                 .titleLabelBottomInset
         } else {
@@ -299,5 +327,25 @@ class SelectLocationViewController: UIViewController, UITableViewDelegate {
             NSLayoutConstraint.activate(tableHeaderFooterViewTopConstraints)
         }
         view.layoutIfNeeded()
+    }
+
+    private func adjustTableViewScrollIndicatorInsets() {
+        guard let tableView = tableView else {
+            return
+        }
+        tableView.verticalScrollIndicatorInsets = tableView.adjustedContentInset
+    }
+}
+
+// MARK: - UISearchBarDelegate
+
+extension SelectLocationViewController: UISearchBarDelegate {
+
+    func searchBarSearchButtonClicked(_ searchBar: UISearchBar) {
+        searchBar.resignFirstResponder()
+    }
+
+    func searchBar(_ searchBar: UISearchBar, textDidChange searchText: String) {
+        dataSource?.filterLocations(by: searchText)
     }
 }